I added a couple of new components: - ThingOfTheMonth - ThingsOfTheMonth - ViewMoreHeader - GameOverviews - TopFreeGames
But I think that some of them might disappear in the future, because I see repetetive code, so I will try to figure out how to get rid of it and create a universal component for list of cards with title and show more button. I already started to do it in the ViewMoreList file, but I stopped it for now and will continue later.

Also I respectively created sections on the HomePage for the components I created above